feat: finalize main page
