+++ src/styles/animations.css
@@ -1,0 +1,11 @@
+@keyframes headerFloat {
+  0%, 100% { transform: translateY(0px) scale(1); }
+  25% { transform: translateY(-5px) scale(1.02); }
+  50% { transform: translateY(-8px) scale(1); }
+  75% { transform: translateY(-3px) scale(1.01); }
+}
+
+@keyframes waveShimmer {
+  0% { opacity: 0.6; transform: scale(1) translateY(0px); }
+  100% { opacity: 0.9; transform: scale(1.1) translateY(-2px); }
+}
